:recycle: ルーティングロジックをリファクタリングしてモジュール化
