+++ src/routes.rs
@@ -1,0 +1,1 @@
+pub mod plans;
